Patch small errors

+++ contrib/modular-transport/helper/mt-queue.h
@@ -18,7 +18,7 @@ namespace ns3 {
 
     public:
       queue_t();
-      queue_t(int lower_limit, int upper_limit, bool (*drop_policy) (event));
+      queue_t(long unsigned int lower_limit, long unsigned int upper_limit, bool (*drop_policy) (event));
       ~queue_t();
       bool pause();
       bool unpause();
